logo_restoration_pipeline/requirements.txt
ultralytics>=8.3.0
google-genai>=0.3.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
Pillow>=10.3.0
python-dotenv
